Move `sdk/datasets/v1` to `sdk/v1/datasets`
